Moved the existing message log files to an ng1 folder
I'll put the new components in the messageLog root folder so we won't need to move them when we get rid of angular 1 from the components library.
