+++ source/components/messageLog/ng1/editableMessageLog.tests.ts
