+++ source/components/messageLog/ng1/editableMessageLog.ts
