+++ source/components/messageLog/ng1/editedByPopover.html
