+++ source/components/messageLog/ng1/messageLog.module.ts
@@ -1,9 +1,9 @@
 import * as angular from 'angular';
 
 import { downgrade as utilitiesDowngrade } from 'typescript-angular-utilities';
-import { moduleName as componentsDowngradeModule } from '../../componentsDowngrade';
+import { moduleName as componentsDowngradeModule } from '../../../componentsDowngrade';
 
-import { moduleName as jqueryModuleName } from '../../services/jquery/jquery.service';
+import { moduleName as jqueryModuleName } from '../../../services/jquery/jquery.service';
 
 import { factoryName, messageLogFactory } from './messageLog.service';
 import { controllerName, directiveName, messageLog, MessageLogController, DeletePermissions, EditPermissions } from './messageLog.directive';
@@ -14,10 +14,10 @@ import {
 	controllerName as editableMessageLogControllerName
 } from './editableMessageLog';
 
-import * as componentServices from '../../services/services.module';
+import * as componentServices from '../../../services/services.module';
 import __dialog = componentServices.dialog;
 
-import { moduleName as templateLoaderModule } from '../../services/templateLoader/templateLoader.service';
+import { moduleName as templateLoaderModule } from '../../../services/templateLoader/templateLoader.service';
 
 export * from './messageLog.service';
 export * from './messageLog.directive';
