+++ source/components/messageLog/ng1/messageLog.service.ts
