+++ source/components/messageLog/ng1/messageLogEditDialog.html
